Restructure theme anatomy (EventFahrplanTheme, ColorSchemes).
+ Add Theme object to allow nested composables to access color values.
+ Docs: https://developer.android.com/develop/ui/compose/designsystems/anatomy
